ui: recorder ui refactor: organize some reused widgets into files
